Add Play Button
I added the play button in a button container, since I want it
and the strict button to be in the same context. I made it a
simple styled component since it shouldn't have to own any
state (or even props) of it's own. All it has to do is pass
the click event to it's parent to start the game.