Finish the game (yeah, I got carried away and forgot to commit)
Add restart button and display. Start and restart buttons behave
as expected. The display shows the number of steps in the current
sequence by default, the string '!!!!!' if the player makes a
mistake, or the string 'WINNER' when the player wins.

I also increased some of the timeouts to make the pace slower.
Unfortunately, I introduced a regression so the audio files aren't
loading. I'm going to go figure that out now.